Refactor gallery logic and improve UI/UX
- Extracted mouse-tracking logic into a custom `useMouseGallery` hook.
- Centralized image sources in `src/constants/images.ts`.
- Added random rotation to images for a more organic feel.
- Improved CSS with better transitions, shadows, and borders.
- Added accessibility attributes (aria-hidden) to decorative images.
- Configured trail length and distance threshold for better feel.
